+++ dist-extension/service-worker-loader.js
@@ -1,1 +1,1 @@
-import './assets/background.ts-ByWT4Gp8.js';
+import './assets/background.ts-DxnWLQa_.js';